Enhance forecasts, backtesting & UI error handling
Multiple improvements across forecasting, evaluation, UI and persistence:

- API/routes: extract predicted_quantity from Forecast objects; proactively refresh alerts when DB has none; adjust chart date range to last 90 days with a 7-day future extension; compute sales revenue using per-product weighted average price; aggregate forecasts by product and date and backfill historical gaps using a scale from overlap with actuals.
- Backtesting: add _generate_backtest_forecasts to generate on-the-fly historical forecasts (trained up to period start) and use it as a fallback when stored forecasts are missing; treat NULL generated_at as legacy forecasts in daily/weekly endpoints.
- Forecast evaluator: restrict to daily-level forecasts, exclude fake sales, and switch from plain MAPE to WMAPE (weighted) for more robust accuracy metrics; use predicted_quantity field.
- Model trainer: set generated_at on forecasts and batch upsert snapshot saves to reduce DB round-trips and handle errors with rollback.
- Models/utils: ensure MAE/RMSE return floats and use forward-fill reindexing more idiomatically.
- Frontend: improve fetch error handling in app.js (display server error message), enable chart spanGaps, and include CSRF header + better response handling in settings restore flow.

Overall goals: improve robustness, accuracy, backtesting support, and performance (fewer DB calls), and provide clearer error handling in the UI.

+++ my app/models/arima.py
@@ -50,8 +50,8 @@ def forecast_arima(db_conn, product_id, days_ahead=7, order=(5,1,0)):
         # 3. Calculate model performance metrics (on training data)
         fitted_values = model_fit.fittedvalues
         residuals = ts['y'] - fitted_values
-        mae = np.mean(np.abs(residuals))
-        rmse = np.sqrt(np.mean(residuals**2))
+        mae = float(np.mean(np.abs(residuals)))
+        rmse = float(np.sqrt(np.mean(residuals**2)))
 
         # 4. Format results
         future_dates = create_future_dates(ts, days_ahead)
